backend/requirements.txt
streamlit==1.45.0
pandas==2.2.3
plotly==6.0.1
pydeck==0.9.1
python-dotenv==1.1.0
openai==1.82.0
fastapi==0.115.0
uvicorn==0.30.0
